requirements.txt
python-dotenv
python-binance
pandas
numpy
ta-lib
tensorflow
scikit-learn
stable-baselines3
gymnasium
requests
